+++ CORE/include/common.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <iostream>
+#include <string>
+#include <vector>
+#include <filesystem>
+#include <fstream>
+#include <exception>
+#include "Bridge.hpp"
+
+namespace fs = std::filesystem;
+
+extern Bridge bridge;