+++ CORE/include/utils.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "common.hpp"
+
+std::vector<std::string> split(const std::string& text);